+++ test/object-store/thread_safe_reference.cpp
@@ -884,6 +884,73 @@ TEST_CASE("thread safe reference") {
         }
     }
 
+    SECTION("create TSR to pre-existing objects inside write transaction") {
+        auto create_ref = [&](auto&& fn) -> ThreadSafeReference {
+            ThreadSafeReference ref;
+            {
+                SharedRealm r2 = Realm::get_shared_realm(config);
+                r2->begin_transaction();
+                auto obj = fn(r2);
+                r2->commit_transaction();
+                r2->begin_transaction();
+                ref = obj;
+                r2->commit_transaction();
+            };
+            return ref;
+        };
+
+        SECTION("object") {
+            auto obj = create_ref([](auto& r) {
+                           return create_object(r, "int object", {{"value", INT64_C(7)}});
+                       }).resolve<Object>(r);
+            REQUIRE(obj.is_valid());
+            REQUIRE(obj.get_column_value<int64_t>("value") == 7);
+        }
+
+        SECTION("object list") {
+            auto list = create_ref([](auto& r) {
+                            auto obj = create_object(r, "int array object",
+                                                     {{"value", AnyVector{AnyDict{{"value", INT64_C(0)}}}}});
+                            return List(r, obj.obj(), get_table(*r, "int array object")->get_column_key("value"));
+                        }).resolve<List>(r);
+            REQUIRE(list.is_valid());
+            REQUIRE(list.size() == 1);
+        }
+
+        SECTION("int list") {
+            auto list = create_ref([](auto& r) {
+                            auto obj = create_object(r, "int array", {{"value", AnyVector{{INT64_C(1)}}}});
+                            return List(r, obj.obj(), get_table(*r, "int array")->get_column_key("value"));
+                        }).resolve<List>(r);
+            REQUIRE(list.is_valid());
+            REQUIRE(list.size() == 1);
+        }
+
+        SECTION("object results") {
+            auto results =
+                create_ref([](auto& r) {
+                    auto obj =
+                        create_object(r, "int array object", {{"value", AnyVector{AnyDict{{"value", INT64_C(0)}}}}});
+                    Results results = List(r, obj.obj(), get_table(*r, "int array object")->get_column_key("value"))
+                                          .sort({{"value", true}});
+                    REQUIRE(results.size() == 1);
+                    return results;
+                }).resolve<Results>(r);
+            REQUIRE(results.is_valid());
+            REQUIRE(results.size() == 1);
+        }
+
+        SECTION("int results") {
+            auto results = create_ref([](auto& r) {
+                               auto obj = create_object(r, "int array", {{"value", AnyVector{{INT64_C(1)}}}});
+                               return List(r, obj.obj(), get_table(*r, "int array")->get_column_key("value"))
+                                   .sort({{"self", true}});
+                           }).resolve<Results>(r);
+            REQUIRE(results.is_valid());
+            REQUIRE(results.size() == 1);
+        }
+    }
+
     SECTION("create TSR inside cancelled write transaction") {
         auto create_ref = [&](auto&& fn) -> ThreadSafeReference {
             ThreadSafeReference ref;