+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 * ThreadSafeReference no longer pins the source transaction version for anything other than a Results backed by a Query.
+* A ThreadSafeReference to a Results backed by a collection can now be created inside a write transaction as long as the collection was not created in the current write transaction.
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)